pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "Personal blog"
authors = ["Balthazar Rouberol <br@imap.cc>"]

[tool.poetry.dependencies]
python = "^3.9"
pelican = "^4.6.0"
Markdown = "^3.3.4"
Pygments = "^2.9.0"
Pillow = "^9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."pygments.lexers"]
"bash" = "bashlexer:ExtendedBashLexer"
